Move UI components to organized directories
- Created and populated media directory with avatar and image components
- Moved overlay components (dropdown-menu, context-menu, hover-card, popover, sheet)
- Added proper README documentation for each component
- Created consistent barrel exports (index.ts files)
- Added deprecation notices to original component locations
- Updated COMPONENT-MOVEMENT-PLAN.md with progress tracking
