+++ src/components/ui/media/avatar/index.ts
@@ -1,0 +1,2 @@
+export * from './avatar';
+export * from './avatar-placeholder';
