+++ src/components/ui/media/image/index.ts
@@ -1,0 +1,2 @@
+export * from './optimized-image';
+export * from './desktop-optimized-image';
