+++ src/components/ui/media/index.ts
@@ -1,0 +1,2 @@
+export * from './avatar';
+export * from './image';
