+++ src/components/ui/overlays/index.ts
@@ -1,3 +1,8 @@
 export * from './dialog';
 export * from './glow-dialog';
 export * from './alert-dialog';
+export * from './hover-card';
+export * from './popover';
+export * from './dropdown-menu';
+export * from './context-menu';
+export * from './sheet';
